requirements.txt
# Web scraping
scrapy==2.12.0
selenium==4.28.1
pdfplumber

# Data analysis
pandas==2.2.3
nltk
gensim
sentence-transformers

# Database
SQLAlchemy
pgvector
psycopg2-binary

# Development tools
ipykernel>=6.21.2
jupyterlab>=3.6.1

# Debugging Utilities
requests==2.31.0 
ipdb==0.13.13

# Environment configuration
python-dotenv
